+ tag-image table

+++ ImagePocket/ImagePocket/ImageRepository.swift
@@ -12,6 +12,8 @@ import SQLite
 final class ImageRepository {
     
     private let _table = Table("Image")
+    private let _tagImageTable = Table("TagImage")
+    
     public static let instance = ImageRepository()
     
     func createTable() throws {
@@ -19,6 +21,7 @@ final class ImageRepository {
         let tableQuery = _table.create(ifNotExists: true) { t in
             t.column(Columns.id, primaryKey: true)
             t.column(Columns.localIdentifier)
+            t.column(Columns.creationDate)
         }
         
         let indexQuery = _table.createIndex([Columns.localIdentifier], ifNotExists: true)
@@ -27,19 +30,48 @@ final class ImageRepository {
         try DataStore.instance.db.run(indexQuery)
     }
     
-    func getAll() -> [ImageEntity] {
-        var result = [ImageEntity]()
+    func createTagImageTable() throws {
         
-        if let rows = try? DataStore.instance.db.prepare(_table){
+        let tableQuery = _tagImageTable.create(ifNotExists: true) { t in
+            t.column(TagImageColumns.id, primaryKey: true)
+            t.column(TagImageColumns.imageId)
+            t.column(TagImageColumns.tagId)
+        }
+        
+        let indexQuery = _tagImageTable.createIndex([TagImageColumns.imageId], ifNotExists: true)
+        
+        try DataStore.instance.db.run(tableQuery)
+        try DataStore.instance.db.run(indexQuery)
+    }
+
+    
+    func getAll() -> [ImageEntity] {
+        
+        var result = [Int64: ImageEntity]()
+        
+        let table =  _table.select(_table[Columns.id], _table[Columns.localIdentifier], _table[Columns.creationDate], _tagImageTable[TagImageColumns.tagId])
+                            .join(_tagImageTable, on: TagImageColumns.imageId == _table[Columns.id])
+        
+        if let rows = try? DataStore.instance.db.prepare(table){
             rows.forEach{ row in
-                let item = ImageEntity(id: row[Columns.id], localIdentifier: row[Columns.localIdentifier], creationDate: row[Columns.creationDate])
-                result.append(item)
+                if let item = result[row[Columns.id]] {
+                    item.appendTagId(id: row[TagImageColumns.tagId])
+                }
+                else {
+                    let item = ImageEntity(id: row[Columns.id], localIdentifier: row[Columns.localIdentifier], creationDate: row[Columns.creationDate])
+                    item.appendTagId(id: row[TagImageColumns.tagId])
+                    result[item.id] = item
+                }
             }
         }
-        return result
+        
+        return result.values.toArray()
     }
     
     func remove(_ entities: [ImageEntity]) -> Void {
+        if entities.isEmpty {
+            return
+        }
         _ = _table.filter(entities.map {$0.id}.contains(Columns.id)).delete()
     }
     
@@ -56,6 +88,10 @@ final class ImageRepository {
         }
         remove(forRemove)
         addOrUpdate(forAddOrUpdate)
+        
+        entities.forEach{entity in
+            addOrUpdateTagImage(imageId: entity.id, entity.tags)
+        }
     }
     
     private func addOrUpdate(_ entities: [ImageEntity]) -> Void {
@@ -69,10 +105,22 @@ final class ImageRepository {
         }
     }
     
+    private func addOrUpdateTagImage(imageId: Int64, _ entities: [TagEntity]){
+        entities.forEach{entity in
+            let query = _tagImageTable.insert(TagImageColumns.imageId <- imageId, TagImageColumns.tagId <- entity.id)
+            let _ = try? DataStore.instance.db.run(query)
+        }
+    }
+    
     private struct Columns {
         static let id = Expression<Int64>("id")
         static let localIdentifier = Expression<String>("localIdentifier")
         static let creationDate = Expression<Date?>("creationDate")
     }
     
+    private struct TagImageColumns {
+        static let id = Expression<Int64>("id")
+        static let imageId = Expression<Int64>("imageId")
+        static let tagId = Expression<Int64>("tagId")
+    }
 }
